Migrate `LocalTimes` to Java Time

+++ time/src/main/java/io/spine/time/DtPreconditions.java
@@ -44,7 +44,7 @@ class DtPreconditions {
     /**
      * Ensures that the passed value is not null and the delta value is positive.
      */
-    static void checkArguments(Object value, int delta) {
+    static void checkArguments(Object value, long delta) {
         checkNotNull(value);
         checkPositive(delta);
     }
